requirements.txt
celery
cryptography
email_validator
Faker
Flask
flask_sqlalchemy
Flask-Babel
flask-caching
Flask-Compress
Flask-Limiter
Flask-Login
Flask-Mail
Flask-Migrate
flask-restx
Flask-Session
Flask-Talisman
Flask-WTF
itsdangerous
jsonschema
lxml
numpy>=1.24.0
openpyxl
pandas>=2.0.0
pingouin
plotly
Pylint
pymysql
pytest
python-dotenv
python-magic
python-magic-bin; sys_platform == 'win32'
redis
requests
scipy>=1.11.0
simpleeval
SQLAlchemy
statsmodels
waitress
Werkzeug
wtforms_sqlalchemy
xlsxwriter
lifelines
mkdocs
mkdocs-material
deepdiff
rich
gunicorn
asteval
defusedxml
